# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Movie streaming platform API — enterprise modular monolith
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /subscriptions/{id}/cancel:
    post:
      operationId: cancelSubscription
      tags:
        - subscriptions
      summary: Cancel a user subscription
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Subscription cancelled
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Subscription'
...
  /users:
    get:
      operationId: listUsers
      tags:
        - users
      summary: List users with advanced filtering
      description: |
        Get paginated list of users with advanced filtering options.
        Supports search, status, role, date range, and sorting.
        Requires `users:read` permission.
        Data is cached for 5 minutes.
      security:
        - bearerAuth: []
      parameters:
        - name: search
          in: query
          schema:
            type: string
          description: Search by username, first name, last name, telegram ID, or email
        - name: status
          in: query
          schema:
            type: string
            enum:
              - active
              - inactive
              - blocked
              - deleted
              - suspended
          description: Filter by user status
        - name: role
          in: query
          schema:
            type: string
            enum:
              - superadmin
              - admin
              - manager
              - moderator
              - user
              - viewer
          description: Filter by user role
        - name: isActive
          in: query
          schema:
            type: boolean
          description: Filter by active status
        - name: isBlocked
          in: query
          schema:
            type: boolean
          description: Filter by blocked status
        - name: isDeleted
          in: query
          schema:
            type: boolean
          description: Filter by deleted status
        - name: startDate
          in: query
          schema:
            type: string
            format: date-time
          description: Filter users created after this date (ISO 8601)
        - name: endDate
          in: query
          schema:
            type: string
            format: date-time
          description: Filter users created before this date (ISO 8601)
        - name: sortBy
          in: query
          schema:
            type: string
            enum:
              - createdAt
              - updatedAt
              - lastLoginAt
              - username
            default: createdAt
          description: Sort field
        - name: sortOrder
          in: query
          schema:
            type: string
            enum:
              - asc
              - desc
            default: desc
          description: Sort order
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
            default: 1
          description: Page number
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
          description: Number of users per page
        - name: includeDeleted
          in: query
          schema:
            type: boolean
            default: false
          description: Include soft-deleted users
      responses:
        '200':
          description: Users retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserListResponse'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden - Insufficient permissions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '429':
          description: Rate limit exceeded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /users/{id}:
    get:
      operationId: getUser
      tags:
        - users
      summary: Get user by ID
      description: |
        Get detailed user information including profile and stats.
        Requires `users:read` permission.
        Data is cached for 5 minutes.
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: User ID
        - name: includeDeleted
          in: query
          schema:
            type: boolean
            default: false
          description: Include soft-deleted user
      responses:
        '200':
          description: User retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserDetailResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    patch:
      operationId: updateUser
      tags:
        - users
      summary: Update user
      description: |
        Update user profile, status, or role.
        Requires `users:update` permission.
        Cannot update own role or status if it would lock you out.
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: User ID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateUserInput'
      responses:
        '200':
          description: User updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserDetailResponse'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    delete:
      operationId: deleteUser
      tags:
        - users
      summary: Delete user
      description: |
        Soft delete or permanently delete a user.
        Requires `users:delete` permission.
        Cannot delete yourself.
        Soft-deleted users can be restored.
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: User ID
        - name: soft
          in: query
          schema:
            type: boolean
            default: true
          description: Soft delete (true) or permanent delete (false)
      responses:
        '200':
          description: User deleted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: User soft deleted successfully
                  timestamp:
                    type: string
                    format: date-time
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /users/{id}/block:
    post:
      operationId: blockUser
      tags:
        - users
      summary: Block or unblock user
      description: |
        Block or unblock a user. Blocked users cannot access the platform.
        Requires `users:block` permission.
        Cannot block yourself.
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: User ID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BlockUserInput'
      responses:
        '200':
          description: User blocked/unblocked successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserDetailResponse'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /users/export:
    post:
      operationId: exportUsers
      tags:
        - users
      summary: Export users
      description: |
        Export users in JSON or CSV format.
        Requires `users:export` permission.
        Maximum export limit is 10,000 users.
      security:
        - bearerAuth: []
      parameters:
        - name: format
          in: query
          schema:
            type: string
            enum:
              - json
              - csv
            default: json
          description: Export format
        - name: search
          in: query
          schema:
            type: string
          description: Filter by search query
        - name: status
          in: query
          schema:
            type: string
            enum:
              - active
              - inactive
              - blocked
              - deleted
              - suspended
          description: Filter by user status
        - name: role
          in: query
          schema:
            type: string
            enum:
              - superadmin
              - admin
              - manager
              - moderator
              - user
              - viewer
          description: Filter by user role
        - name: startDate
          in: query
          schema:
            type: string
            format: date-time
          description: Filter users created after this date
        - name: endDate
          in: query
          schema:
            type: string
            format: date-time
          description: Filter users created before this date
      responses:
        '200':
          description: Export successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExportResponse'
            application/csv:
              schema:
                type: string
                description: CSV file content
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '429':
          description: Rate limit exceeded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  schemas:
...
    Subscription:
      type: object
      properties:
        id:
          type: string
        userId:
          type: string
        planId:
          type: string
        status:
          type: string
        startDate:
          type: string
          format: date-time
        endDate:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
...
    BlockUserInput:
      type: object
      required:
        - blocked
      properties:
        blocked:
          type: boolean
          description: True to block, false to unblock
        reason:
          type: string
          nullable: true
          description: Reason for blocking
...
    UserListResponse:
      type: object
      required:
        - success
        - data
        - meta
        - timestamp
      properties:
        success:
          type: boolean
          example: true
        data:
          type: array
          items:
            $ref: '#/components/schemas/User'
        meta:
          type: object
          required:
            - total
            - page
            - limit
            - totalPages
            - hasNext
            - hasPrevious
          properties:
            total:
              type: integer
              description: Total number of users
              example: 15234
            page:
              type: integer
              description: Current page
              example: 1
            limit:
              type: integer
              description: Items per page
              example: 20
            totalPages:
              type: integer
              description: Total pages
              example: 762
            hasNext:
              type: boolean
              description: Whether there is a next page
              example: true
            hasPrevious:
              type: boolean
              description: Whether there is a previous page
              example: false
            cached:
              type: boolean
              description: Whether response came from cache
              example: false
            cachedAt:
              type: string
              nullable: true
              format: date-time
              description: Cache timestamp
        timestamp:
          type: string
          format: date-time
          description: Response timestamp
    UserDetailResponse:
      type: object
      required:
        - success
        - data
        - timestamp
      properties:
        success:
          type: boolean
          example: true
        data:
          $ref: '#/components/schemas/UserDetail'
        timestamp:
          type: string
          format: date-time
...
    ExportResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          type: object
          properties:
            format:
              type: string
              description: Export format
              example: json
            count:
              type: integer
              description: Number of exported users
              example: 1000
            filename:
              type: string
              description: Export filename
              example: users-export-2026-01-15.json
        timestamp:
          type: string
          format: date-time
    ErrorResponse:
      type: object
      required:
        - success
        - error
      properties:
        success:
          type: boolean
          example: false
        error:
          type: object
          required:
            - code
            - message
            - timestamp
          properties:
            code:
              type: string
              description: Error code
              example: VALIDATION_ERROR
            message:
              type: string
              description: Error message
              example: Invalid input data
            details:
              type: array
              nullable: true
              items:
                type: object
                properties:
                  field:
                    type: string
                    description: Field name
                  message:
                    type: string
                    description: Field error message
              description: Detailed error information
            timestamp:
              type: string
              format: date-time
              description: Error timestamp
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT authentication token